+++ CHANGELOG.md
@@ -31,3 +31,4 @@
 - Added account region endpoint
 - Added account user endpoint
 - Added terms of service endpoint
+- Added terms of service state endpoint